src/requirements.txt
amqp==5.1.1
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2
django-allauth==0.56.0
django-bootstrap-v5==1.0.11
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-polymorphic==3.1.0
django-resized==1.0.2
djangorestframework==3.14.0
easy-thumbnails==2.8.5
gunicorn==21.2.0
idna==3.4
kombu==5.3.2
lxml==4.9.3
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.0
reportlab==4.0.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
